Move restutils.py into http.py
